requirements.txt
transformers==4.35.0
torch==2.1.0
networkx==3.2
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
pyyaml==6.0.1
tqdm==4.66.1
